fix: move some logic to prevent circular import
