+++ shell/app/modules/org/stores/announcement.ts
@@ -24,8 +24,6 @@ import {
   unPublishAnnouncement,
 } from '../services/announcement';
 import { getDefaultPaging } from 'common/utils';
-import { eventHub } from 'common/utils/event-hub';
-import orgStore from 'app/org-home/stores/org';
 
 interface INotice extends IPagingResp<ORG_ANNOUNCEMENT.Item> {
   publishedList: ORG_ANNOUNCEMENT.Item[];
@@ -42,18 +40,6 @@ const initState: INotice = {
 const announcementStore = createStore({
   name: 'announcement',
   state: initState,
-  subscriptions() {
-    eventHub.once('layout/mount', () => {
-      const loginUser = userStore.getState((s) => s.loginUser);
-      const orgId = orgStore.getState((s) => s.currentOrg.id);
-      // 非系统管理员
-      if (!loginUser.isSysAdmin && orgId) {
-        announcementStore.effects.getAllNoticeListByStatus('published').then((list) => {
-          layoutStore.reducers.setAnnouncementList(list);
-        });
-      }
-    });
-  },
   effects: {
     async getAnnouncementList({ call, update }, payload: ORG_ANNOUNCEMENT.QueryList) {
       const { list, total } = await call(getAnnouncementList, payload, { paging: { key: 'noticePaging' } });
